feat: adding build.gradle.kts initially to each module with an example java class to be found by the java plugin
Just creating a java file will not supply java tool support since we did not gave the component any meaning yet. To make it into a java-library, we can apply a gradle plugin. Than the folder structure and java file will be recognized.
E.g. "src/main/java" will than be known to be the java sources, and display a proper java package, instead of a plain folder structure